Core: Make Load/Store use 64bit vaddr

+++ Source/Project64-core/N64System/Interpreter/InterpreterOps.h
@@ -243,10 +243,10 @@ protected:
     static Func Jump_CoP1_W[64];
     static Func Jump_CoP1_L[64];
 
-    static void GenerateAddressErrorException(uint32_t VAddr, bool FromRead);
+    static void GenerateAddressErrorException(uint64_t VAddr, bool FromRead);
     static void GenerateOverflowException(void);
-    static void GenerateTLBReadException(uint32_t VAddr, const char * function);
-    static void GenerateTLBWriteException(uint32_t VAddr, const char * function);
+    static void GenerateTLBReadException(uint64_t VAddr, const char * function);
+    static void GenerateTLBWriteException(uint64_t VAddr, const char * function);
 
     static const uint32_t SWL_MASK[4], SWR_MASK[4], LWL_MASK[4], LWR_MASK[4];
     static const int32_t SWL_SHIFT[4], SWR_SHIFT[4], LWL_SHIFT[4], LWR_SHIFT[4];
